fvModels, fvConstraints: Prevent crash when applied to solver-type function objects
This fixes a crash that occured when the field that the fvModel applies
to does not exist at the point of the fvModel's construction. This is
the case for function objects that solve equations, e.g., the
scalarTransport or age functions, as the relevant fields are constructed
on demand at the point of the functions' execution. Sources and
constraints now work correctly for these sorts of equations.